.gitignore add terraform lock file
Although it is best practice to include this as per the HashCorp
documentation, I am excluding it so that it can be initialized on other
machines without issue. I am working on an ARM-based machine and am
unsure on which other devices this may be ran on. You can specify other
types of machines, but as this is public I can't account for all of
them.